Update _file.html.erb
We have setup a redmine for internal use. At the same time, a proxy is setup too so that external users can enjoy our bug tracking service. If the url is created as a full path, external user can't download the file as it is returning the file of internal server address. Having the relative path generated can make both internal and external parties possible to download the file.