Remove ERB::Util#h
The method ERB::Util#h is an alias for ERB::Util#html_escape. It is
applied on a user name whithin a collect condition in DmsfUploadHelper module.

This implementation was also applied in other modules and has
caused HTML status 500 errors wherefore it was removed.

Therefore, it should also precautionary eliminated in DmsfUploadHelper module.