Fixes Column 'dmsf_mail_notification' cannot be null
The partial app/views/dmsf_state/_user_pref.html.erb uses a condition
to include the email_notify setting only if the corresponding
notified_event is enabled. The database field email_notify, however,
does not accept a null value. But this is what occurs when the notified
event is disabled.

This patch changes the DmsfStateController to assign params[:email_notify]
only if the notified_event for dmsf_legacy_notifications is enabled.